LayoutTests/http/tests/security/cross-frame-access-object-setPrototypeOf-expected.txt

CONSOLE MESSAGE: line 22: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
This tests that you can't set the prototype of the window or history objects cross-origin using Object.setPrototypeOf().

PASS: targetWindow instanceof Array should be 'false' and is.
PASS: targetWindow instanceof Array should be 'false' and is.
